Move display function to module directory
